+++ src/Event/HasCommand.php
@@ -2,22 +2,20 @@
 
 namespace League\Tactician\CommandEvents\Event;
 
-use League\Tactician\Command;
-
 /**
  * Holds the command object for an event
  */
 trait HasCommand
 {
     /**
-     * @var Command
+     * @var object
      */
     protected $command;
 
     /**
      * Returns the command
      *
-     * @return Command
+     * @return object
      */
     public function getCommand()
     {