+++ stubs/Command.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace League\Tactician\CommandEvents\Stub;
+
+class Command
+{
+
+}